requirements.txt
requests
python-dotenv
pyTelegramBotAPI
beautifulsoup4
